Move WooPay integration test to separate file to fix CI OOM
The upe-styles test file runs close to the CI worker memory limit.
Adding the getAppearance('woopay_shortcode_checkout', true) integration
test to the same file caused the worker to be killed (SIGTERM).

Split the WooPay-specific tests into woopay-appearance.test.js so they
run in a separate Jest worker process:
- Integration test verifying WooPay rules and link color priority
- Selector routing test for woopay_shortcode_checkout
